+++ tests/XRoadLib.Tests/ProducerDefinitionTest.fs
@@ -2,6 +2,7 @@
 
 open FsUnit
 open NUnit.Framework
+open System
 open System.IO
 open System.Linq
 open System.Xml.Linq
@@ -9,6 +10,8 @@ open XRoadLib
 
 [<TestFixture>]
 module ProducerDefinitionTest =
+    let contractAssembly = typeof<XRoadLib.Tests.Contract.Class1>.Assembly
+
     let xn nm = XName.Get(nm)
     let wsdl nm = XName.Get(nm, NamespaceHelper.WSDL)
     let soap nm = XName.Get(nm, NamespaceHelper.SOAP)
@@ -27,14 +30,14 @@ module ProducerDefinitionTest =
         let service = root.Elements(wsdl "service").SingleOrDefault()
         service |> should not' (be Null)
         service.Attributes().Count() |> should equal 1
-        service.Attribute(xn "name") |> attributeValueShouldEqual "testService"
+        service.Attribute(xn "name") |> attributeValueShouldEqual "TestService"
         service.Elements().Count() |> should equal 1
 
         let port = service.Elements(wsdl "port").SingleOrDefault()
         port |> should not' (be Null)
         port.Attributes().Count() |> should equal 2
-        port.Attribute(xn "name") |> attributeValueShouldEqual "testPort"
-        port.Attribute(xn "binding") |> attributeValueShouldEqual "testBinding"
+        port.Attribute(xn "name") |> attributeValueShouldEqual "TestPort"
+        port.Attribute(xn "binding") |> attributeValueShouldEqual "TestBinding"
 
         let okValues = [soap "address"; xrdns "address"; xrdns "title"]
         port.Elements() |> Seq.filter (fun e -> okValues |> List.exists((=) e.Name) |> not) |> should be Empty
@@ -54,20 +57,20 @@ module ProducerDefinitionTest =
 
     [<Test>]
     let ``empty service description`` () =
-        let doc = ProducerDefinition(XRoadProtocol.Version31, "test", 1u) |> getDocument
+        let doc = ProducerDefinition(contractAssembly, XRoadProtocol.Version31, Nullable(1u)) |> getDocument
         let port = shouldMatchInCommonParts doc xrd
 
         let address = port.Elements(xrd "address").SingleOrDefault()
         address |> should not' (be Null)
         address.IsEmpty |> should equal true
         address.Attributes().Count() |> should equal 1
-        address.Attribute(xn "producer") |> attributeValueShouldEqual "test"
+        address.Attribute(xn "producer") |> attributeValueShouldEqual "test-producer"
 
-        port.Elements(soap "address").Single().Attribute(xn "location") |> attributeValueShouldEqual ""
+        port.Elements(soap "address").Single().Attribute(xn "location") |> attributeValueShouldEqual "http://TURVASERVER/cgi-bin/consumer_proxy"
 
     [<Test>]
     let ``empty legacy format service description`` () =
-        let doc = ProducerDefinition(XRoadProtocol.Version20, "test", 1u) |> getDocument
+        let doc = ProducerDefinition(contractAssembly, XRoadProtocol.Version20, Nullable(1u), "test") |> getDocument
         let port = shouldMatchInCommonParts doc xtee
 
         let address = port.Elements(xtee "address").SingleOrDefault()
@@ -76,12 +79,12 @@ module ProducerDefinitionTest =
         address.Attributes().Count() |> should equal 1
         address.Attribute(xn "producer") |> attributeValueShouldEqual "test"
 
-        port.Elements(soap "address").Single().Attribute(xn "location") |> attributeValueShouldEqual ""
+        port.Elements(soap "address").Single().Attribute(xn "location") |> attributeValueShouldEqual "http://TURVASERVER/cgi-bin/consumer_proxy"
 
     [<Test>]
     let ``should define service location if given`` () =
         let url = "http://securityserveruri"
-        let definition = ProducerDefinition(XRoadProtocol.Version31, "test", 1u)
+        let definition = ProducerDefinition(contractAssembly, XRoadProtocol.Version31, Nullable(1u))
         definition.set_Location(url)
         let doc = definition |> getDocument
         let port = shouldMatchInCommonParts doc xrd
@@ -89,52 +92,56 @@ module ProducerDefinitionTest =
 
     [<Test>]
     let ``should define service title`` () =
-        let title1 = "Ilma keeleta palun"
-        let title2 = "Portugalikeelne loba ..."
-
-        let definition = ProducerDefinition(XRoadProtocol.Version31, "test", 1u)
-        definition.Title.Add("", title1)
-        definition.Title.Add("pt", title2)
+        let definition = ProducerDefinition(contractAssembly, XRoadProtocol.Version31, Nullable(1u))
 
         let doc = definition |> getDocument
         let port = shouldMatchInCommonParts doc xrd
 
         let titleElements = port.Elements(xrd "title") |> List.ofSeq
-        titleElements.Length |> should equal 2
+        titleElements.Length |> should equal 4
 
         let yesCode, noCode = titleElements |> List.partition (fun x -> x.Attributes().Any())
 
-        yesCode.Length |> should equal 1
-        yesCode.Head |> should not' (be Null)
-        yesCode.Head.Value |> should equal title2
-        yesCode.Head.Attribute(xml "lang") |> attributeValueShouldEqual "pt"
+        yesCode.Length |> should equal 3
+        match yesCode with
+        | [ en; et; pt ] ->
+            yesCode |> List.iter (fun x -> x |> should not' (be Null))
+            en.Value |> should equal "XRoadLib test producer"
+            en.Attribute(xml "lang") |> attributeValueShouldEqual "en"
+            et.Value |> should equal "XRoadLib test andmekogu"
+            et.Attribute(xml "lang") |> attributeValueShouldEqual "et"
+            pt.Value |> should equal "Portugalikeelne loba ..."
+            pt.Attribute(xml "lang") |> attributeValueShouldEqual "pt"
+        | _ -> failwith "never"
 
         noCode.Length |> should equal 1
         noCode.Head |> should not' (be Null)
-        noCode.Head.Value |> should equal title1
+        noCode.Head.Value |> should equal "Ilma keeleta palun"
 
     [<Test>]
     let ``can define service title for legacy service`` () =
-        let title1 = "Ilma keeleta palun"
-        let title2 = "Portugalikeelne loba ..."
-
-        let definition = ProducerDefinition(XRoadProtocol.Version20, "test", 1u)
-        definition.Title.Add("", title1)
-        definition.Title.Add("pt", title2)
+        let definition = ProducerDefinition(contractAssembly, XRoadProtocol.Version20, Nullable(1u))
 
         let doc = definition |> getDocument
         let port = shouldMatchInCommonParts doc xtee
 
         let titleElements = port.Elements(xtee "title") |> List.ofSeq
-        titleElements.Length |> should equal 2
+        titleElements.Length |> should equal 4
 
         let yesCode, noCode = titleElements |> List.partition (fun x -> x.Attributes().Any())
 
-        yesCode.Length |> should equal 1
-        yesCode.Head |> should not' (be Null)
-        yesCode.Head.Value |> should equal title2
-        yesCode.Head.Attribute(xml "lang") |> attributeValueShouldEqual "pt"
+        yesCode.Length |> should equal 3
+        match yesCode with
+        | [ en; et; pt ] ->
+            yesCode |> List.iter (fun x -> x |> should not' (be Null))
+            en.Value |> should equal "XRoadLib test producer"
+            en.Attribute(xml "lang") |> attributeValueShouldEqual "en"
+            et.Value |> should equal "XRoadLib test andmekogu"
+            et.Attribute(xml "lang") |> attributeValueShouldEqual "et"
+            pt.Value |> should equal "Portugalikeelne loba ..."
+            pt.Attribute(xml "lang") |> attributeValueShouldEqual "pt"
+        | _ -> failwith "never"
 
         noCode.Length |> should equal 1
         noCode.Head |> should not' (be Null)
-        noCode.Head.Value |> should equal title1
+        noCode.Head.Value |> should equal "Ilma keeleta palun"